Split large file healthchecks.go => healthcheck.go,healthchecks.go
- Adds some pretty printing to the backendconfig hc options
